helm: add credential-default Secret template
Towards https://github.com/giantswarm/giantswarm/issues/6557

This can't be deployed with aws-operator or azure-operator chart anymore
as we want to be able to install multiple releases of the same operator at
the same time. That would case conflict when creating credential-default
Secret which for helm is a resource with the same name.

After some discussions we decided to put this here as credentiald also
creates other credential secret. And this one had even
`giantswarm.io/managed-by: credentiald` label.